Regenerate tests for fix for bug 713
The fix for bug #713 changes the way capital gains as well as basis costs
are shown.  All of these tests use the same input file, which includes
capital gains and losses.  The tests are too long to update and verify
manually but I verified some files and believe the output is correct.
I've therefore updated the tests with the output of the fixed ledger.

Let's take baseline/opt-lot-dates.test as one example.  In the past,
ledger generated the following incorrect Equity entries:

             190.42G  Equity
             236.83G    Capital Gains
             -46.41G    Capital Losses

If you compare the end of the diff of the test, you'll see:

  -                                                                          35.42G
  +                                                                        -155.00G

The difference is: 35.42 - -155.00 = 190.42

This matches exactly the Equity entries previously generated, which
shouldn't be there.